chore: fix CI checks for /stop improvements
Revert the manual 1.69.7→1.69.8 bump in package.json/package-lock.json
(versions are managed by changesets), and extract shared /stop test
fixtures into tests/test-telegram-stop-helpers.mjs so jscpd duplication
stays under the 11% threshold (10.94% after the refactor).
