+++ tests/test-issue-1783-stop-improvements.mjs
@@ -20,24 +20,13 @@
  */
 
 import { isStopTargetRequester, updateQueueCardForCancellation, registerStartStopCommands } from '../src/telegram-start-stop-command.lib.mjs';
+import { makeAsserts, makeStubBot, makeStopCtx as makeCtx, makeStopOptions as makeOptions } from './test-telegram-stop-helpers.mjs';
 
 console.log('='.repeat(80));
 console.log('Unit Tests: /stop command improvements (Issue #1783)');
 console.log('='.repeat(80));
 
-let passed = 0;
-let failed = 0;
-
-function assert(cond, name, details) {
-  if (cond) {
-    console.log(`  ✅ ${name}`);
-    passed++;
-  } else {
-    console.log(`  ❌ ${name}`);
-    if (details !== undefined) console.log(`     ${JSON.stringify(details)}`);
-    failed++;
-  }
-}
+const { counts, assert } = makeAsserts();
 
 const ISSUE_URL = 'https://github.com/link-assistant/hive-mind/issues/1783';
 const UUID = '4d934f71-4cdb-4b8c-b474-582116d12c12';
@@ -131,62 +120,6 @@ console.log('\n--- updateQueueCardForCancellation() ---');
 // Dispatcher integration: /stop <url> updates the queue card on cancel
 // ---------------------------------------------------------------------------
 console.log('\n--- /stop <url> updates the queue card on cancel-queued ---');
-
-function makeStubBot() {
-  const handlers = {};
-  return {
-    handlers,
-    command(name, fn) {
-      handlers[name] = fn;
-    },
-    on() {},
-  };
-}
-
-function makeCtx({ text = '/stop', repliedTo = null, chatType = 'private', chatId = -1001, fromId = REQUESTER_ID, fromUsername = 'requester' } = {}) {
-  const replies = [];
-  const edits = [];
-  return {
-    replies,
-    edits,
-    chat: { id: chatId, type: chatType },
-    from: { id: fromId, username: fromUsername, first_name: 'Test' },
-    message: {
-      message_id: 7,
-      text,
-      date: Math.floor(Date.now() / 1000),
-      reply_to_message: repliedTo,
-    },
-    reply: async (text, opts) => {
-      const r = { chat: { id: chatId }, message_id: 100 + replies.length, text, opts };
-      replies.push(r);
-      return r;
-    },
-    telegram: {
-      getChatMember: async () => ({ status: 'creator' }),
-      editMessageText: async (cId, mId, _x, newText, opts) => {
-        edits.push({ chatId: cId, messageId: mId, text: newText, opts });
-        return true;
-      },
-    },
-  };
-}
-
-function makeOptions(overrides = {}) {
-  return {
-    VERBOSE: false,
-    isOldMessage: () => false,
-    isForwardedOrReply: () => false,
-    isGroupChat: () => true,
-    isChatAuthorized: () => true,
-    isTopicAuthorized: () => true,
-    buildAuthErrorMessage: () => '❌ not authorized',
-    addBreadcrumb: () => {},
-    isChatStopped: () => false,
-    getStoppedChatRejectMessage: () => '❌ stopped',
-    ...overrides,
-  };
-}
 
 function makeQueueItem({ id, url, tool, sessionName = null, requesterUserId = null, ctx = null, messageInfo = null }) {
   return { id, url, tool, sessionName, requesterUserId, ctx, messageInfo };
@@ -487,8 +420,8 @@ console.log('\n--- /stop <UUID>: requester can stop their own session in a group
 
 // -------------------------- summary ----------------------------------------
 console.log('\n' + '='.repeat(80));
-console.log(`Result: ${passed} passed, ${failed} failed`);
+console.log(`Result: ${counts.passed} passed, ${counts.failed} failed`);
 console.log('='.repeat(80));
-if (failed > 0) {
+if (counts.failed > 0) {
   process.exit(1);
 }
