+++ tests/test-issue-524-stop-uuid.mjs
@@ -15,37 +15,13 @@
  */
 
 import { extractStopSessionId, registerStartStopCommands } from '../src/telegram-start-stop-command.lib.mjs';
+import { makeAsserts, makeStubBot, makeStopCtx as makeCtx, makeStopOptions as makeOptions } from './test-telegram-stop-helpers.mjs';
 
 console.log('='.repeat(80));
 console.log('Unit Tests: /stop <UUID> command (Issue #524)');
 console.log('='.repeat(80));
 
-let passed = 0;
-let failed = 0;
-
-function assert(cond, name, details) {
-  if (cond) {
-    console.log(`  ✅ ${name}`);
-    passed++;
-  } else {
-    console.log(`  ❌ ${name}`);
-    if (details !== undefined) console.log(`     ${JSON.stringify(details)}`);
-    failed++;
-  }
-}
-
-function assertEqual(actual, expected, name) {
-  const pass = JSON.stringify(actual) === JSON.stringify(expected);
-  if (pass) {
-    console.log(`  ✅ ${name}`);
-    passed++;
-  } else {
-    console.log(`  ❌ ${name}`);
-    console.log(`     expected: ${JSON.stringify(expected)}`);
-    console.log(`     actual:   ${JSON.stringify(actual)}`);
-    failed++;
-  }
-}
+const { counts, assert, assertEqual } = makeAsserts();
 
 // ----- extractStopSessionId --------------------------------------------------
 console.log('\n--- extractStopSessionId() ---');
@@ -74,62 +50,6 @@ assertEqual(extractStopSessionId(null, undefined), { sessionId: null, source: nu
 
 // ----- registerStartStopCommands /stop dispatcher ----------------------------
 console.log('\n--- registerStartStopCommands() /stop dispatcher ---');
-
-function makeStubBot() {
-  const handlers = {};
-  return {
-    handlers,
-    command(name, fn) {
-      handlers[name] = fn;
-    },
-    on() {},
-  };
-}
-
-function makeCtx({ text = '/stop', repliedTo = null, chatType = 'private', chatId = -1001, fromId = 42 } = {}) {
-  const replies = [];
-  const edits = [];
-  return {
-    replies,
-    edits,
-    chat: { id: chatId, type: chatType },
-    from: { id: fromId, username: 'tester', first_name: 'Test' },
-    message: {
-      message_id: 7,
-      text,
-      date: Math.floor(Date.now() / 1000),
-      reply_to_message: repliedTo,
-    },
-    reply: async (text, opts) => {
-      const r = { chat: { id: chatId }, message_id: 100 + replies.length, text, opts };
-      replies.push(r);
-      return r;
-    },
-    telegram: {
-      getChatMember: async () => ({ status: 'creator' }),
-      editMessageText: async (cId, mId, _x, newText, opts) => {
-        edits.push({ chatId: cId, messageId: mId, text: newText, opts });
-        return true;
-      },
-    },
-  };
-}
-
-function makeOptions(overrides = {}) {
-  return {
-    VERBOSE: false,
-    isOldMessage: () => false,
-    isForwardedOrReply: () => false,
-    isGroupChat: () => true,
-    isChatAuthorized: () => true,
-    isTopicAuthorized: () => true,
-    buildAuthErrorMessage: () => '❌ not authorized',
-    addBreadcrumb: () => {},
-    isChatStopped: () => false,
-    getStoppedChatRejectMessage: () => '❌ stopped',
-    ...overrides,
-  };
-}
 
 // Test 1: /stop <UUID> in private chat invokes stopIsolatedSession.
 {
@@ -257,8 +177,8 @@ function makeOptions(overrides = {}) {
 
 // -------------------------- summary ------------------------------------------
 console.log('\n' + '='.repeat(80));
-console.log(`Result: ${passed} passed, ${failed} failed`);
+console.log(`Result: ${counts.passed} passed, ${counts.failed} failed`);
 console.log('='.repeat(80));
-if (failed > 0) {
+if (counts.failed > 0) {
   process.exit(1);
 }
